Implement market pro orders
We would like to avoid having market orders as they have proven
to be a risk for the broker. The next best thing for the client is a
limit order with a percent_through_market, which denotes how deeply
into the other side of the book a user wants to place a limit order.