Switch architect-py to uv (from poetry) (#995)
# Description

I got rid of dynamic versioning, in favor of a pre-commit script that
checks the two places that need to be in sync. You have to install the
pre-commit stuff via `uvx pre-commit install`. This will be followed up
with a GitHub Actions based publishing workflow which will also run the
check. I verified that `uv build` gets the correct version set.

## Changelog

- [ ] Update `CHANGELOG.md` to reflect the changes in this PR