Split up Start/Resume Algos as Separate Functions (#1289)
# Description

## Motivation
I wanted it to be possible to have an `on_start` function for the algos,
such that the algo would start up and then immediately do a task.
It was not possible to do this.

so there were functions dealing with starting from a paused state, I
renamed those `resume`.


fixed a minor race issue with
`load_algo`, where an algo would resume and then immediately stopped, it
might run `start_tasks` after and then not be able to cancel them


## Note

technically, it is possible to combine `on_start` and `on_resume` within
each algo, but just to reduce ambiguity, I propose just keeping them
separate


## Changelog

- [X] Update `CHANGELOG.md` to reflect the changes in this PR

---------

Co-authored-by: Andrew Lee <126016409+andrew-lee-architect@users.noreply.github.com>
Co-authored-by: Megan Mason <megan@architect.co>
Co-authored-by: Andrew Lee <665282+chairmanlee8@users.noreply.github.com>
Co-authored-by: kevin-jasieniecki-architect <kevin@architect.co>
Co-authored-by: Brett Harrison <brett@architect.gg>
Co-authored-by: Brett Harrison <116864397+brett-harrison-architect@users.noreply.github.com>