recently out algo orders (#1308)
# Description

- feat: OpenAlgoOrders to AlgoExecutor
- improvement: AlgoExecutor now stores recently_out_algo_orders, and
get_algo_order now queries it on top of working orders
- fix: spreader stopping before child outs acked

# Dev Notes
- SDK still handles a) query open/roo algo orders then b) historical
orders. This can be modified to be handled by the algo service. It was
omitted for now in this PR because it is a little more involved and less
urgent than its corresponding change in the oms service.
- Retry loop for canceling roo algo orders DNE but will follow once it
is implemented for roo orders in the folio

## Changelog

- [x] Update `CHANGELOG.md` to reflect the changes in this PR